gan/requirements.txt
absl-py==0.14.1
anyio==3.3.2
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
clang==5.0
click==8.0.2
colorama==0.4.4
cryptography==3.3.2
cursor==1.3.4
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
halo==0.0.28
idna==3.2
ipaddress==1.0.23
Jinja2==3.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
log-symbols==0.0.14
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.2
pip-chill==1.0.1
protobuf==3.18.1
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.2
pywin32==301
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
sentry-sdk==1.4.3
six==1.15.0
smmap==4.0.0
sniffio==1.2.0
spell==0.44.2
spinners==0.0.24
starlette==0.16.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
